+++ ChangeLog.md
@@ -15,3 +15,4 @@
 
 - MqClient.unconsume change to async with no return
 - RpcCodec.encodeResponse remove class info
+- fix SpringServiceBootstrap ssl config bug